refactor(action): cleanup linux and coreboot builds
- linux kernel and coreboot no longer use the same build function to
  reduce complexity and improve readability
- blob handling for coreboot is not much simpler and more straight
  forward
- linux, coreboot and edk2 specific functions (like corebootOpts) are
  moved out from recipes.go into their respective file (like coreboot.go)
